requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
